pyproject.toml
[project]
name = "jobflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (==3.15.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "aioschedule (>=0.5.2,<0.6.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "markdown (>=3.10,<4.0)",
    "yookassa (>=3.9.0,<4.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.125.0,<0.126.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
